plaid controller
