+++ package.json
@@ -2,9 +2,9 @@
   "name": "Indow",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "backend/app.js",
   "scripts": {
-    "start": "nodemon backend/server.js",
+    "start": "nodemon -L backend/app.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
